refactor(cli): Move completion.py to cli/services
