+++ src/dstack/_internal/cli/services/completion.py
